# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestion de boutique",
        "description": "Documentation de l'API REST pour la gestion de boutique (catégories, produits, acheteurs, achats). CCP 2026.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Serveur de développement local"
        }
    ],
    "paths": {
        "/acheteurs/{id}/acheter": {
            "post": {
                "tags": [
                    "Achats"
                ],
                "summary": "Enregistrer un nouvel achat pour un acheteur",
                "operationId": "e2054940650e1c24d1729ca475694081",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'acheteur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "produit_id",
                                    "quantite",
                                    "date_achat"
                                ],
                                "properties": {
                                    "produit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantite": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "date_achat": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-28"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Achat enregistré"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Acheteur introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/acheteurs": {
            "get": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Lister tous les acheteurs",
                "operationId": "3048d16a74d411acda3cbc100611d321",
                "responses": {
                    "200": {
                        "description": "Liste des acheteurs"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Créer un nouvel acheteur",
                "operationId": "f6a1f76165e42643d45d60f141ce6fa6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Fatou Diop"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "fatou.diop@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "771234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Acheteur créé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé (réservé gestionnaire/admin)"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/acheteurs/{id}": {
            "get": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Afficher le détail d'un acheteur (avec son historique d'achats)",
                "operationId": "09860c055857cc04361e86f7c71d5d68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de l'acheteur"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "404": {
                        "description": "Acheteur introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Modifier un acheteur",
                "operationId": "7711ae9ab9c69a6b4b86d82d763f187c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "email"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Fatou Diop Ndiaye"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "fatou.ndiaye@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "771234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Acheteur modifié"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Acheteur introuvable"
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Acheteurs"
                ],
                "summary": "Supprimer un acheteur",
                "operationId": "9045802096d4fba8184cb073107e0eee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Acheteur supprimé"
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Acheteur introuvable"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Entrez le token obtenu via /login, sans le mot 'Bearer' devant (Swagger l'ajoute automatiquement).",
                "bearerFormat": "Sanctum token",
                "scheme": "bearer"
            }
        }
    },
...
}